requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
mysql-connector-python==8.0.33
passlib==1.7.4
bcrypt>=4.0.0
requests==2.31.0
python-dotenv==1.0.0
